requirements.txt
fastapi
uvicorn
pydantic
pdfplumber 
python-docx